requirements.txt
streamlit
pandas
matplotlib
python-dotenv
requests
streamlit
pandas

groq
duckdb
phi
agno
